+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/views.cljs
@@ -1,5 +1,6 @@
 (ns sixsq.nuvla.ui.edge-detail.views
   (:require
+    [cljs.pprint :refer [cl-format pprint]]
     [clojure.string :as str]
     [re-frame.core :refer [dispatch subscribe]]
     [reagent.core :as r]
@@ -233,16 +234,83 @@
 (defn Load
   [resources]
   [uix/Accordion
-   (let [load-stats (u/load-statistics resources)]
-     [plot/HorizontalBar {:height  50
-                          :data    {:labels   (map :label load-stats)
-                                    :datasets [{:data (map :percentage load-stats)}]}
-                          :options {:scales {:xAxes [{:type  "linear"
-                                                      :ticks {:beginAtZero true
-                                                              :max         100}}]
-                                             :yAxes [{:gridLines {:display false}}]}}}])
-   :label "Load Percentages"
-   :icon "thermometer half"])
+   (let [load-stats (u/load-statistics resources)
+         number-of-stats (count load-stats)]
+     ; TODO: if the number-of-stats grows if should split into a new row
+     [ui/Grid {:columns   number-of-stats,
+               :stackable true
+               :divided   true}
+
+      [ui/GridRow
+       (for [stat load-stats]
+         [ui/GridColumn
+
+          [:div
+           [plot/Doughnut {:height  250
+                           :data    {:labels    (:label stat)
+                                     :datasets  [{:data            [(:percentage stat), (:value stat)]
+                                                  :backgroundColor [
+                                                                    "rgb(230, 99, 100)",
+                                                                    "rgba(155, 99, 132, 0.1)",
+                                                                    "rgb(230, 99, 100)"
+                                                                    ]
+                                                  :borderColor     ["rgba(230, 99, 100,1)"]
+                                                  :borderWidth     3}]}
+                           :options {:legend   {:display true
+                                                :labels  {
+                                                          :fontColor "grey"
+                                                          }}
+                                     :title    {:display   true
+                                                :text      (:title stat)
+                                                :position  "bottom"}
+                                     :maintainAspectRatio false
+                                     :circumference       4.14
+                                     :rotation            -3.64
+                                     :cutoutPercentage    60}}]]
+
+
+          [ui/Container {:key         (:topic stat)
+                         :text-align  :center}
+           [ui/LabelGroup {:key   (:topic stat)
+                           :size  "tiny"}
+            [ui/Label {:color "blue"
+                       :basic true
+                       :image true}
+             "Topic: "
+             [ui/LabelDetail
+              (first (:data-gateway stat))]]
+            [ui/Label {:color "blue"
+                       :basic true
+                       :image true}
+             "Raw sample: "
+             [ui/LabelDetail
+              (last (:data-gateway stat))]]]
+           ]
+
+
+          ; TODO: the data-gateway stats should be in a popup instead of raw text. But fails some unknown reason,
+          ;[ui/Popup
+          ; {:trigger        (r/as-element [ui/Icon {:name "info circle"}])
+          ;  :content        "Let your NuvlaBox apps subscribe to the internal MQTT to access these values locally"
+          ;  :header         "data-gateway"
+          ;  :position       "right center"
+          ;  :inverted       true
+          ;  :wide           true
+          ;
+          ;  :on             "hover"
+          ;  :hide-on-scroll true}]
+          ])]])
+   :label [:span "Resource Consumption " [ui/Popup
+                                          {:trigger        (r/as-element [ui/Icon {:name "info circle"}])
+                                           :content        "Let your NuvlaBox apps subscribe to the internal MQTT topics
+                                          to access these values locally"
+                                           :header         "data-gateway"
+                                           :position       "right center"
+                                           :inverted       true
+                                           :wide           true
+                                           :on             "hover"
+                                           :hide-on-scroll true}]]
+   :icon  "thermometer half"])
 
 
 (defn StatusSection